src/requirements.txt
Flask
Flask-RESTful
flask-swagger-ui
gunicorn
python-dotenv
sqlalchemy
requests
cryptography
Flask-cors
